增加Int64

+++ parser/const/number.cpp
@@ -1,10 +1,11 @@
 #include "number.h"
-
+#include "int32.h"
+#include "int64.h"
 namespace Parser {
     std::unique_ptr<AST::ExprAST> ParseNumber(Module *module) {
         auto token = module->nowToken;
         if (token->type == token_int) {
-            return ParseInt32(module);
+            return ParseInt64(module);
         }
         return nullptr;
     }
